Remove registrar and registration contract, and add TangleID description
Registrar maintains a registration contract recording the participants'
DID document address and public key. It is used as a look-up table to
retreive entities' public key and DID.
Since the public key is included on DID document, the registration contract
can be replaced with the address of DID document. Thus, registrar can
also be removed, and TangleID section is added back to the Component section.